+++ R/req-perform.R
@@ -87,7 +87,6 @@ req_perform <- function(
   }
 
   req <- req_verbosity(req, verbosity)
-  req <- auth_sign(req)
 
   req <- cache_pre_fetch(req, path)
   if (is_response(req)) {
@@ -111,19 +110,7 @@ req_perform <- function(
     sys_sleep(delay, "for retry backoff")
     n <- n + 1
 
-    resp <- tryCatch(
-      req_perform1(req, path = path, handle = handle),
-      error = function(err) {
-        error_cnd(
-          message = "Failed to perform HTTP request.",
-          class = c("httr2_failure", "httr2_error"),
-          parent = err,
-          request = req,
-          call = error_call,
-          trace = trace_back()
-        )
-      }
-    )
+    resp <- req_perform1(req, path = path, handle = handle)
     req_completed(req_prep)
 
     if (retry_is_transient(req, resp)) {
@@ -132,8 +119,7 @@ req_perform <- function(
       signal(class = "httr2_retry", tries = tries, delay = delay)
     } else if (!reauth && resp_is_invalid_oauth_token(req, resp)) {
       reauth <- TRUE
-      req <- auth_sign(req, TRUE)
-      req_prep <- req_prepare(req)
+      req_prep <- req_prepare(req, reauth = TRUE)
       handle <- req_handle(req_prep)
       delay <- 0
     } else {
@@ -154,13 +140,32 @@ handle_resp <- function(req, resp, error_call = caller_env()) {
   }
 
   if (is_error(resp)) {
+    resp$request <- req
+    resp$call <- error_call
     cnd_signal(resp)
   } else if (error_is_error(req, resp)) {
-    body <- error_body(req, resp, error_call)
-    resp_abort(resp, req, body, call = error_call)
+    cnd <- resp_failure_cnd(req, resp, error_call = error_call)
+    cnd_signal(cnd)
   } else {
     resp
   }
+}
+
+resp_failure_cnd <- function(req, resp, error_call = caller_env()) {
+  status <- resp_status(resp)
+  desc <- resp_status_desc(resp)
+  message <- glue("HTTP {status} {desc}.")
+
+  info <- error_body(req, resp, error_call)
+
+  catch_cnd(abort(
+    c(message, resp_auth_message(resp), info),
+    status = status,
+    resp = resp,
+    class = c(glue("httr2_http_{status}"), "httr2_http", "httr2_error", "rlang_error"),
+    request = req,
+    call = error_call
+  ))
 }
 
 req_perform1 <- function(req, path = NULL, handle = NULL) {
@@ -168,28 +173,31 @@ req_perform1 <- function(req, path = NULL, handle = NULL) {
   the$last_response <- NULL
   signal(class = "httr2_perform")
 
-  if (!is.null(path)) {
-    res <- curl::curl_fetch_disk(req$url, path, handle)
-    body <- new_path(path)
-  } else {
-    res <- curl::curl_fetch_memory(req$url, handle)
-    body <- res$content
+  err <- capture_curl_error({
+    fetch <- curl_fetch(handle, req$url, path)
+  })
+  if (is_error(err)) {
+    return(err)
   }
 
   # Ensure cookies are saved to disk now, not when request is finalised
   curl::handle_setopt(handle, cookielist = "FLUSH")
   curl::handle_setopt(handle, cookiefile = NULL, cookiejar = NULL)
 
-  resp <- new_response(
-    method = req_method_get(req),
-    url = res$url,
-    status_code = res$status_code,
-    headers = as_headers(res$headers),
-    body = body,
-    request = req
-  )
-  the$last_response <- resp
-  resp
+  the$last_response <- create_response(req, fetch$curl_data, fetch$body)
+  the$last_response
+}
+
+curl_fetch <- function(handle, url, path) {
+  if (!is.null(path)) {
+    curl_data <- curl::curl_fetch_disk(url, path, handle)
+    body <- new_path(path)
+  } else {
+    curl_data <- curl::curl_fetch_memory(url, handle)
+    body <- curl_data$content
+  }
+
+  list(curl_data = curl_data, body = body)
 }
 
 req_verbosity <- function(req, verbosity, error_call = caller_env()) {
@@ -229,19 +237,20 @@ last_request <- function() {
 }
 
 # Must call req_prepare(), then req_handle(), then after the request has been
-# performed, req_completed()
-req_prepare <- function(req) {
+# performed, req_completed() (on the prepared requests)
+req_prepare <- function(req, reauth = FALSE) {
+  req <- auth_sign(req, reauth = reauth)
   req <- req_method_apply(req)
   req <- req_body_apply(req)
-
+  req
+}
+req_handle <- function(req) {
   if (!has_name(req$options, "useragent")) {
     req <- req_user_agent(req)
   }
 
-  req
-}
-req_handle <- function(req) {
   handle <- curl::new_handle()
+  curl::handle_setopt(handle, url = req$url)
   curl::handle_setheaders(handle, .list = headers_flatten(req$headers))
   curl::handle_setopt(handle, .list = req$options)
   if (length(req$fields) > 0) {